+++ app/javascript/parsePageContent.jsx
@@ -9,26 +9,12 @@ import SignInButton from './Authentication/SignInButton';
 import SignOutButton from './Authentication/SignOutButton';
 import SignUpButton from './Authentication/SignUpButton';
 import Spoiler from './Spoiler';
-import { lazyWithBundleHashCheck } from './checkBundleHash';
-
-const EventAdminMenu = lazyWithBundleHashCheck(() => import(/* webpackChunkName: "events-app" */ './EventsApp/EventPage/EventAdminMenu'));
-const LongFormEventDetails = lazyWithBundleHashCheck(() => import(/* webpackChunkName: "events-app" */ './EventsApp/EventPage/LongFormEventDetails'));
-const ProposeEventButton = lazyWithBundleHashCheck(() => import(/* webpackChunkName: 'propose-event-button' */ './EventProposals/ProposeEventButton'));
-const RunsSection = lazyWithBundleHashCheck(() => import(/* webpackChunkName: "events-app" */ './EventsApp/EventPage/RunsSection'));
-const ShortFormEventDetails = lazyWithBundleHashCheck(() => import(/* webpackChunkName: "events-app" */ './EventsApp/EventPage/ShortFormEventDetails'));
-const WithdrawMySignupButton = lazyWithBundleHashCheck(() => import(/* webpackChunkName: 'withdraw-my-signup-button' */ './EventsApp/EventPage/WithdrawMySignupButton'));
 
 export const DEFAULT_COMPONENT_MAP = {
-  EventAdminMenu,
-  LongFormEventDetails,
-  ProposeEventButton,
-  RunsSection,
-  ShortFormEventDetails,
   Spoiler,
   SignInButton,
   SignOutButton,
   SignUpButton,
-  WithdrawMySignupButton,
 };
 
 // adapted from html-to-react library
